span_links: make v04 and v05 tests consistent (#2168)
* span_links: make v04 and v05 tests consistent

* lint

* fix test

* typo
